move bpmn-semantic-utils.ts to test/shared
