+++ test/unit/component/registry/bpmn-model-registry.test.ts
@@ -16,7 +16,7 @@ limitations under the License.
 
 import type { EdgeBpmnSemantic, ShapeBpmnSemantic } from '@lib/component/registry';
 import { BpmnModelRegistry } from '@lib/component/registry/bpmn-model-registry';
-import { expectAssociationFlow, expectLane, expectMessageFlow, expectPool, expectSequenceFlow, expectStartEvent } from '../../helpers/bpmn-semantic-utils';
+import { expectAssociationFlow, expectLane, expectMessageFlow, expectPool, expectSequenceFlow, expectStartEvent } from '@test/shared/model/bpmn-semantic-utils';
 import { associationFlowInModel, laneInModel, messageFlowInModel, poolInModel, sequenceFlowInModel, startEventInModel } from '../../helpers/bpmn-model-utils';
 
 const bpmnModelRegistry = new BpmnModelRegistry();
